Renaming `user` to `resident` for clarity
