+++ components/eligibility/index.tsx
@@ -5,7 +5,7 @@ import { useStore } from "react-redux"
 
 export default function EligibilityOutcome(): JSX.Element {
   const store = useStore<Store>()
-  const isEligible = store.getState().user.isEligible
+  const isEligible = store.getState().resident.isEligible
 
   return isEligible ? <Eligible /> : <NotEligible />
 }
