+++ lib/store/index.ts
@@ -1,17 +1,17 @@
-import user from "./user"
-import { User } from "../types/user"
+import resident from "./resident"
+import { Resident } from "../types/resident"
 import { createStore, combineReducers } from "redux"
 import { createWrapper, Context } from "next-redux-wrapper"
 
 export interface Store {
-  user: User
+  resident: Resident
 }
 
 const reducer = combineReducers({
-  user: user.reducer
+  resident: resident.reducer
 })
 
 // Store function
 const store = (context: Context) => createStore(reducer)
-export const wrapper = createWrapper(store);
+export const wrapper = createWrapper(store)
 export default store
